fix: payments 페이지 구조 변경
